+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file based on the
 
 ### Bugfixes
 - Fixed GeoShapeProvided relation parameter position
+- Function score query: corrected the `score_method` `average` to `avg` #975 
 
 ### Added
 - Connection option to convert JSON bigint results to strings can now be set [#717](https://github.com/ruflin/Elastica/issues/717)